Fix source timing such that -ve times mean not set
It was tempting to equate begin time = end time = 0 to mean that the
times were not set and the whole file should be used (times could be
just booleans).  This is inconsistent with end times being inclusive,
which implies begin time = end time = 0 means "give me one frame at
index 0".

A better convention, which Juicer uses, is to use negative times to
mean not set; at least at the source anyway, where negative times are
meaningless otherwise.  This patches the HTKSource and Source
interface to do that.

Many other sources don't do this yet.