Fixing various timing issues
One was the constness of the TimeStamp() function, which the VADGate
was not doing.  Another was to do with the way TimeOffset() calculated
offsets.  Now all seems well, but I've said that before...

There is still an issue where the end time for a word in the result
structure is short of the total number of frames that should be
accounted for.  I think this is just a remnant of label pushing, but
can't be sure right now.