Rename Signal.{h,cpp} to TracterFPE.{h,cpp}.
Signal.h is not a good name (at least under Mac OS X).
The main reason is that signal.h (with small s) is a standard include file. And
due to the fact that Max OS X is not case sensitive (but only case preserving
by default), this cause some troubles.